Implemented top, bottom, left and right halo exchange.
- TODO: need to implement the corner exchanges. Make sure this is done generalised so it can be done with boxes as well.

+++ src/conway/include/conway.h
@@ -5,7 +5,7 @@
 
 // TODO: comment
 namespace conway {
-class ConwaysArray2DWithHalo : public array2d::Array2DWithHalo<short int> {
+class ConwaysArray2DWithHalo : public array2d::Array2DWithHalo<int> {
    public:
     ConwaysArray2DWithHalo(int n_rows, int n_cols);
